requirements.txt
fastapi
uvicorn
streamlit
requests
google-generativeai
psycopg2-binary
python-dotenv
